// packages/contracts/typechain-types/CorpusRegistry.ts
export declare namespace CorpusRegistry {
  export type PatronConfigStruct = {
    creatorShare: BigNumberish;
    investorShare: BigNumberish;
    treasuryShare: BigNumberish;
    creatorAddr: AddressLike;
    investorAddr: AddressLike;
    treasuryAddr: AddressLike;
  };

  export type PatronConfigStructOutput = [
    creatorShare: bigint,
    investorShare: bigint,
    treasuryShare: bigint,
    creatorAddr: string,
    investorAddr: string,
    treasuryAddr: string
  ] & {
    creatorShare: bigint;
    investorShare: bigint;
    treasuryShare: bigint;
    creatorAddr: string;
    investorAddr: string;
    treasuryAddr: string;
  };

  export type KernelConfigStruct = {
    approvalThreshold: BigNumberish;
    gtmBudget: BigNumberish;
    minPatronPulse: BigNumberish;
  };

  export type KernelConfigStructOutput = [
    approvalThreshold: bigint,
    gtmBudget: bigint,
    minPatronPulse: bigint
  ] & { approvalThreshold: bigint; gtmBudget: bigint; minPatronPulse: bigint };

  export type PulseConfigStruct = {
    hederaTokenAddr: AddressLike;
    totalSupply: BigNumberish;
    priceUsdCents: BigNumberish;
  };

  export type PulseConfigStructOutput = [
    hederaTokenAddr: string,
    totalSupply: bigint,
    priceUsdCents: bigint
  ] & { hederaTokenAddr: string; totalSupply: bigint; priceUsdCents: bigint };

  export type CorpusStruct = {
    id: BigNumberish;
    name: string;
    category: string;
    creator: AddressLike;
    patron: CorpusRegistry.PatronConfigStruct;
    kernel: CorpusRegistry.KernelConfigStruct;
    pulse: CorpusRegistry.PulseConfigStruct;
    createdAt: BigNumberish;
    active: boolean;
  };

  export type CorpusStructOutput = [
    id: bigint,
    name: string,
    category: string,
    creator: string,
    patron: CorpusRegistry.PatronConfigStructOutput,
    kernel: CorpusRegistry.KernelConfigStructOutput,
    pulse: CorpusRegistry.PulseConfigStructOutput,
    createdAt: bigint,
    active: boolean
  ] & {
    id: bigint;
    name: string;
    category: string;
    creator: string;
    patron: CorpusRegistry.PatronConfigStructOutput;
    kernel: CorpusRegistry.KernelConfigStructOutput;
    pulse: CorpusRegistry.PulseConfigStructOutput;
    createdAt: bigint;
    active: boolean;
  };
}