// packages/contracts/typechain-types/CorpusNameService.ts
export interface CorpusNameServiceInterface extends Interface {
  getFunction(
    nameOrSignature:
      | "hasName"
      | "isNameAvailable"
      | "nameOf"
      | "registerName"
      | "registry"
      | "resolveName"
  ): FunctionFragment;

  getEvent(nameOrSignatureOrTopic: "NameRegistered"): EventFragment;

  encodeFunctionData(
    functionFragment: "hasName",
    values: [BigNumberish]
  ): string;
  encodeFunctionData(
    functionFragment: "isNameAvailable",
    values: [string]
  ): string;
  encodeFunctionData(
    functionFragment: "nameOf",
    values: [BigNumberish]
  ): string;
  encodeFunctionData(
    functionFragment: "registerName",
    values: [BigNumberish, string]
  ): string;
  encodeFunctionData(functionFragment: "registry", values?: undefined): string;
  encodeFunctionData(functionFragment: "resolveName", values: [string]): string;

  decodeFunctionResult(functionFragment: "hasName", data: BytesLike): Result;
  decodeFunctionResult(
    functionFragment: "isNameAvailable",
    data: BytesLike
  ): Result;
  decodeFunctionResult(functionFragment: "nameOf", data: BytesLike): Result;
  decodeFunctionResult(
    functionFragment: "registerName",
    data: BytesLike
  ): Result;
  decodeFunctionResult(functionFragment: "registry", data: BytesLike): Result;
  decodeFunctionResult(
    functionFragment: "resolveName",
    data: BytesLike
  ): Result;
}

export namespace NameRegisteredEvent {
  export type InputTuple = [corpusId: BigNumberish, name: string];
  export type OutputTuple = [corpusId: bigint, name: string];
  export interface OutputObject {
    corpusId: bigint;
    name: string;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export interface CorpusNameService extends BaseContract {
  connect(runner?: ContractRunner | null): CorpusNameService;
  waitForDeployment(): Promise<this>;

  interface: CorpusNameServiceInterface;

  queryFilter<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;
  queryFilter<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;

  on<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  on<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  once<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  once<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  listeners<TCEvent extends TypedContractEvent>(
    event: TCEvent
  ): Promise<Array<TypedListener<TCEvent>>>;
  listeners(eventName?: string): Promise<Array<Listener>>;
  removeAllListeners<TCEvent extends TypedContractEvent>(
    event?: TCEvent
  ): Promise<this>;

  hasName: TypedContractMethod<[corpusId: BigNumberish], [boolean], "view">;

  isNameAvailable: TypedContractMethod<[name: string], [boolean], "view">;

  nameOf: TypedContractMethod<[corpusId: BigNumberish], [string], "view">;

  registerName: TypedContractMethod<
    [corpusId: BigNumberish, name: string],
    [void],
    "nonpayable"
  >;

  registry: TypedContractMethod<[], [string], "view">;

  resolveName: TypedContractMethod<[name: string], [bigint], "view">;

  getFunction<T extends ContractMethod = ContractMethod>(
    key: string | FunctionFragment
  ): T;

  getFunction(
    nameOrSignature: "hasName"
  ): TypedContractMethod<[corpusId: BigNumberish], [boolean], "view">;
  getFunction(
    nameOrSignature: "isNameAvailable"
  ): TypedContractMethod<[name: string], [boolean], "view">;
  getFunction(
    nameOrSignature: "nameOf"
  ): TypedContractMethod<[corpusId: BigNumberish], [string], "view">;
  getFunction(
    nameOrSignature: "registerName"
  ): TypedContractMethod<
    [corpusId: BigNumberish, name: string],
    [void],
    "nonpayable"
  >;
  getFunction(
    nameOrSignature: "registry"
  ): TypedContractMethod<[], [string], "view">;
  getFunction(
    nameOrSignature: "resolveName"
  ): TypedContractMethod<[name: string], [bigint], "view">;

  getEvent(
    key: "NameRegistered"
  ): TypedContractEvent<
    NameRegisteredEvent.InputTuple,
    NameRegisteredEvent.OutputTuple,
    NameRegisteredEvent.OutputObject
  >;

  filters: {
    "NameRegistered(uint256,string)": TypedContractEvent<
      NameRegisteredEvent.InputTuple,
      NameRegisteredEvent.OutputTuple,
      NameRegisteredEvent.OutputObject
    >;
    NameRegistered: TypedContractEvent<
      NameRegisteredEvent.InputTuple,
      NameRegisteredEvent.OutputTuple,
      NameRegisteredEvent.OutputObject
    >;
  };
}